Added a Charging Jump animation state and modified PlayerAnimController.cs
NOTE: There isn't actually a ChargingJump animation yet, because it seems like the animations were done in Blender and I don't know how to make animations yet, but I took the "GrindIdleForward" animation which looks close enough and put that as a placeholder for when the player holds down the spacebar to start charging up a jump. The code has comments to explain the changes, but maybe I overcomplicated things by doing it this way.