Addressing Hannah's feedback regarding the debug statements
Addressing the issue of "Delete the debug statements the console is overrun currently". I've commented away some of the debug statements that were previously used in the Player.cs script. This can be un-done again later, but I left in important messages like "The player died by impact" "The player respawned" etc.